重構InstallersPage、AdditionalInstallersPage
